+++ BeyondTrust/beyondtrust_modules/client/__init__.py
@@ -31,6 +31,14 @@ class ApiClient(requests.Session):
             timeout=60,
         )
 
+    def get_vault_activity(self, end_time: int) -> requests.Response:
+        return self.post(
+            f"{self._base_url}/api/reporting",
+            data={"generate_report": "VaultAccountActivity", "duration": 0, "end_time": end_time},
+            headers={"Content-Type": "application/x-www-form-urlencoded"},
+            timeout=60,
+        )
+
     def get_session_listing(self, end_time: int) -> requests.Response:
         return self.post(
             f"{self._base_url}/api/reporting",