+++ MicrosoftEntraID/azure_ad/asset_connector/user_assets.py
@@ -199,7 +199,7 @@ class EntraIDAssetConnector(AsyncAssetConnector):
             time=datetime.timestamp(user.created_date_time) if user.created_date_time is not None else 0,
             metadata=metadata,
             user=user_data,
-            type_name="User Inventory",
+            type_name="User Inventory Info: Collect",
             enrichments=enrichments or None,
         )
         return user_ocsf_model
@@ -211,7 +211,6 @@ class EntraIDAssetConnector(AsyncAssetConnector):
         groups: list[UserOCSFGroup] = []
         try:
             user_groups = await self.client.users.by_user_id(user_id).member_of.get()
-
             if user_groups and user_groups.value:
                 for group in user_groups.value:
                     if isinstance(group, Group):