+++ SentinelOne/sentinelone_module/asset_connector/device_assets.py
@@ -9,7 +9,6 @@ from functools import cached_property
 from typing import Any, Optional
 
 from dateutil.parser import isoparse
-
 from sekoia_automation.asset_connector import AssetConnector
 from sekoia_automation.asset_connector.models.ocsf.base import (
     Metadata,
@@ -33,9 +32,10 @@ from sekoia_automation.asset_connector.models.ocsf.device import (
     RiskLevelStr,
 )
 from sekoia_automation.storage import PersistentJSON
-from sentinelone_module.client import SentinelOneClient
+
 from sentinelone_module.asset_connector.models import SentinelOneAgent
 from sentinelone_module.base import SentinelOneModule
+from sentinelone_module.client import SentinelOneClient
 
 
 class SentinelOneDeviceAssetConnector(AssetConnector):