+++ Sophos/main.py
@@ -1,6 +1,8 @@
+from sophos_module.account_validator import SophosAccountValidator
 from sophos_module.action_sophos_edr_deisolate import ActionSophosEDRDeIsolateEndpoint
 from sophos_module.action_sophos_edr_isolate import ActionSophosEDRIsolateEndpoint
 from sophos_module.action_sophos_edr_run_scan import ActionSophosEDRScan
+from sophos_module.asset_connector import SophosDeviceAssetConnector
 from sophos_module.base import SophosModule
 from sophos_module.trigger_sophos_edr_events import SophosEDREventsTrigger
 from sophos_module.trigger_sophos_xdr_query import SophosXDRIOCQuery
@@ -12,5 +14,7 @@ if __name__ == "__main__":
     module.register(ActionSophosEDRIsolateEndpoint, "sophos_edr_isolate_endpoint")
     module.register(ActionSophosEDRDeIsolateEndpoint, "sophos_edr_deisolate_endpoint")
     module.register(ActionSophosEDRScan, "sophos_edr_run_scan")
+    module.register(SophosDeviceAssetConnector, "sophos_device_asset_connector")
+    module.register(SophosAccountValidator, "sophos_account_validator")
 
     module.run()